chore(config): 更新标准化构建流程和相关配置
