+++ package.json
@@ -21,14 +21,7 @@
     "commit": "git-cz",
     "commitlint": "commitlint --config .commitlintrc.js -e -V",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s",
-    "pub": "node ./scripts/pub.js",
-    "cm": "node ./scripts/commit.js",
-    "release": "./scripts/release.sh"
-  },
-  "husky": {
-    "hooks": {
-      "commit-msg": "commitlint -e $GIT_PARAMS"
-    }
+    "pub": "node ./scripts/publish.js"
   },
   "lint-staged": {
     "src/*.ts": [
@@ -65,7 +58,7 @@
     "@rollup/plugin-eslint": "^8.0.2",
     "@testing-library/dom": "^8.18.1",
     "@testing-library/jest-dom": "^5.16.5",
-    "@types/jest": "^29.1.1",
+    "@types/jest": "^28.1.1",
     "@typescript-eslint/eslint-plugin": "^5.38.1",
     "@typescript-eslint/parser": "^5.38.1",
     "chalk": "4.1.2",
@@ -77,9 +70,9 @@
     "eslint-plugin-prettier": "^4.2.1",
     "husky": "^8.0.1",
     "inquirer": "^8.0.0",
-    "jest": "^29.1.2",
+    "jest": "^28.1.1",
     "jest-axe": "^6.0.0",
-    "jest-environment-jsdom": "^29.1.2",
+    "jest-environment-jsdom": "^28.1.1",
     "jest-report": "^0.1.11",
     "lint-staged": "^12.5.0",
     "prettier": "^2.7.1",
